left-sidebar: Create collapsible private messages section.
This commit changes the method of rendering PMs to DOM.

`_build_private_messages_list` function is modified now to
first fetch all the conversations present and then we convert these
conversation objects in nodes according to the condition whether
PM section is zoomed or not, if not zoomed (which is by-default)
then we slice out only first 4 PMs out of the list and convert them
into nodes and if length of all the conversations is > 4, we push
a node with `more conversations` at the end of this list to DOM
clicking on which zooms in the PM section in which all the
conversations are rendered to the DOM.

This keeps the count of number of unreads present in convos which
are under `more conversations` and passes their count to the html
of `more_pms` as parameter in the function `render_more_pms` of
`more_pms` node.

We now directly convert the conversation objects into nodes handling
the conditions mentioned above while building private messages list in
pm_list and then send it to `pm_list_dom` to finally update the `vdom`,
earlier all these conversations were send to `pm_list_dom` and were then
converted into nodes and renderd to DOM.

In building private messages we also take care of the concept of `active_pm`,
before converting conversations into nodes, we check if the `active_pm` is
present in the first 4 PMs, otherwise we append the `active_pm` separately
into the nodes passed on to `pm_list_dom` taking the count of number of
shown conversations to 5.

Adds a function of `handle_narrow_activated()` which updates the
PM list by pushing the active_pm into the list according to the narrow_filter
passed in it as a parameter and this also adds the class of `active-filter`
to PM section if narrow_filter contains PMs.

Adds a function `handle_narrow_deactivated()` which updates the PMs list
according to the narrow_filter, this also zooms out from the PM section and
removes the class of `active_private_messages_section` if narrow_filer
does not contain `is/pm-with` or `is/private` and resizes the
STREAM LIST accordingly.

Adds 2 on click functions in initialize() of pm_list which takes care of
zooming in/out of PM section when clicked over `more conversations`
button in zoomed out view and `Back` in `zoomed In view.

Adds a icon-button to PM section in left_sidebar.hbs clicking on which
makes the PM section collapse/expand.